refactor(embedding): extract model-lifecycle state machine (#173)
Fowler Extract Class + Replace Type Code with explicit state. Introduce
embedding_model_lifecycle.py:
- ModelState enum names the distinct load outcomes — UNINITIALIZED, LOADED,
  PACKAGE_ABSENT (sentence_transformers not installed), MODEL_FILES_ABSENT
  (package present, weights not cached → download), LOAD_RAISED (construction
  failed). This is the state space where a silent-degradation bug can hide;
  naming it is the seam a second encoder will map each state onto.
- _EmbeddingLifecycleMixin owns device detection/resolution and the load path
  (_detect_device, _resolve_device, _fallback_to_cpu, _ensure_model split into
  _construct_model/_download_model/_load_model, _trigger_background_install).
- DEFAULT_MODEL_REVISION and embedding_cache_dir move here; embedding_engine
  re-exports them for import compatibility.

embedding_engine.py is now the concrete neural provider (encode path + LRU
cache) plus the composition root (get_embedding_engine). Behaviour preserved:
PACKAGE_ABSENT → hash fallback + background install; MODEL_FILES_ABSENT →
download; LOAD_RAISED → propagates — identical to the pre-split loader. Existing
embedding tests pass unchanged.

File sizes: engine 263, lifecycle 294, provider 115 (all <=300); every method
<=40 lines.

Closes #173
